+++ src/components/LayerUI.tsx
@@ -387,15 +387,8 @@ const LayerUI = ({
         elements={elements}
         appState={appState}
         actionManager={actionManager}
-        onExportToBackend={(elements) => {
-          UIOptions.canvasActions.export.onExportToBackend &&
-            UIOptions.canvasActions.export.onExportToBackend(
-              elements,
-              appState,
-              canvas,
-            );
-        }}
         exportOpts={UIOptions.canvasActions.export}
+        canvas={canvas}
       />
     );
   };